+++ src/components/account/FormInput.tsx
@@ -1,6 +1,9 @@
 import styled from '@emotion/styled';
+import { useState } from 'react';
 import type { ComponentProps } from 'react';
 import type { ErrorOption, UseFormRegisterReturn } from 'react-hook-form';
+import HideIcon from 'assets/icons/hide_pw.svg';
+import ShowIcon from 'assets/icons/show_pw.svg';
 
 interface FormInputProps extends ComponentProps<'input'> {
   register: UseFormRegisterReturn;
@@ -9,8 +12,11 @@ interface FormInputProps extends ComponentProps<'input'> {
   isShowLabel?: boolean;
 }
 
+interface ErrorStyleProps {
+  isError: boolean;
+}
+
 const FormInput = ({
-  name,
   type,
   register,
   placeholder,
@@ -18,16 +24,31 @@ const FormInput = ({
   errors,
   isShowLabel = false,
 }: FormInputProps) => {
+  const { name } = register;
+  const isPasswordInput = name.includes('password');
+  const [showPassword, setShowPassword] = useState<boolean>(false);
+
+  const handleOnTogglePassword = () => {
+    setShowPassword((state) => !state);
+  };
+
   return (
     <div>
       {isShowLabel && <Label htmlFor={name}>{label}</Label>}
-      <Input
-        type={type}
-        id={name}
-        placeholder={placeholder}
-        isError={errors !== undefined}
-        {...register}
-      />
+      <InputBox>
+        <Input
+          type={isPasswordInput ? (showPassword ? 'text' : 'password') : type}
+          id={name}
+          placeholder={placeholder}
+          isError={errors !== undefined}
+          {...register}
+        />
+        {isPasswordInput && (
+          <PasswordButton type="button" onClick={handleOnTogglePassword}>
+            {showPassword ? <ShowIcon /> : <HideIcon />}
+          </PasswordButton>
+        )}
+      </InputBox>
       <ErrorText>{errors?.message}</ErrorText>
     </div>
   );
@@ -40,7 +61,11 @@ const Label = styled.label`
   color: ${({ theme }) => theme.colors.gray_01};
 `;
 
-const Input = styled.input<{ isError: boolean }>`
+const InputBox = styled.div`
+  position: relative;
+`;
+
+const Input = styled.input<ErrorStyleProps>`
   width: 100%;
   padding: 8px 0 6px;
   border-bottom: 1px solid
@@ -65,4 +90,10 @@ const ErrorText = styled.p`
   margin-top: 8px;
   color: ${({ theme }) => theme.colors.error};
   ${({ theme }) => theme.fonts.body_09};
+`;
+
+const PasswordButton = styled.button`
+  position: absolute;
+  bottom: 6px;
+  right: 0;
 `;
